Add a parser for use in tests
Inspired matklad's article "How to Test" [1], add a parser so that
tests can be written very concisely.  For example:

assert_staff_snapshot!("### f# |");

To give the parser something to parse into, add a "model" module,
which contains the data structures that the parser will
produce.  (These same data structures will eventually be what users of
this library will also be using.)

The "model" types differ from the "input" types in that they exist at
the "music" level, rather than the "engraving" level.  For example, a
middle-C quarter note in a staff with a treble clef will be
represented in the "model" types as a note with a pitch of C4, but in
the "input" types will be represented as a note located at staff space
-1.  "model" and "input" do both use the same representation of note
durations, however.

[1]: https://matklad.github.io/2021/05/31/how-to-test.html